// apps/main/src/models/v2/iam/orgMembership.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 */
import type { SubjectType } from './subjectType';

/**
 * A user membership from the perspective of an org
 */
export interface OrgMembership {
  /** The membership id */
  id: string;
  /** The time the membership was created */
  created_at: string;
  /** The user id of the membership */
  user_id: string;
  /** The display name of the user membership */
  user_display_name: string;
  /** The primary email address of the user membership if available */
  user_primary_email_address?: string;
  /** The membership subject type. */
  subject_type: SubjectType;
  /** The membership subject identity. */
  subject: string;
  /** The membership scope if applicable composed by <resource-kind>:<resource-id>. It will be empty for organization level memberships. */
  scope?: string;
}